backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.10
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.6.2
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.19.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.64
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langsmith==0.3.45
llama-cloud==0.1.23
llama-cloud-services==0.6.30
llama-index==0.12.40
llama-index-agent-openai==0.4.9
llama-index-cli==0.4.3
llama-index-core==0.12.40
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.4
llama-index-llms-openai==0.4.3
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.30
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.84.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
stripe==12.2.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0
